Issue 7707: Fix DMA address truncation from uint64_t to uint32_t

+++ components/devices/bgq/mu2/model/CollectiveMulticastDmaModel.h
@@ -107,8 +107,8 @@ namespace PAMI
 								  classroute);
 	      }
 	      else {
-		unsigned src_pa;
-		unsigned dst_pa;
+		uint64_t src_pa;
+		uint64_t dst_pa;
 		
 		if (isroot) {
 		  Kernel_MemoryRegion_t memRegion;
